리팩토링: 기존의 RacingCarGame클래스에서 자동차 이름 저장하는 기능을 분리하여 SaveCarNames 클래스에 저장
