+++ index.html
@@ -28,6 +28,6 @@
         <div id="racing-result"></div>
       </div>
     </div>
-    <script type="module" src="src/index.js"></script>
+    <script type="module" src="src/racing-car-game.js"></script>
   </body>
 </html>
