+++ src/car.js
@@ -1,4 +1,4 @@
-export default class car {
+export default class Car {
     constructor(name) {
         this.name = name;
         this.distance = 0;
